feat(react-chemdoodleweb:sketcher): :sparkles: adds sketcher canvas
moves sketcher and viewer to their own files
